rename + change to stage manager
